+++ CustomerService.Data/Services/ServiceModels/CustomerServiceModel.cs
@@ -8,10 +8,10 @@ namespace CustomerService.Data.Services.ServiceModels
 {
     public class CustomerServiceModel
     {
-        public decimal ID { get; set; }
+        public long ID { get; set; }
         public string Name { get; set; }
         public string ContactEmail { get; set; }
-        public decimal? MobileNo { get; set; }
+        public int? MobileNo { get; set; }
         public List<TransactionServiceModel> Transactions { get; set; } = new List<TransactionServiceModel>();
     }
 }
